[GLI-43-44] Add menu information in persistence (#51)

+++ reservation-api/src/main/java/ca/ulaval/glo4002/reservation/domain/fullcourse/recipe/allergy/AllergyMain.java
@@ -1,13 +1,11 @@
 package ca.ulaval.glo4002.reservation.domain.fullcourse.recipe.allergy;
 
-import java.math.BigDecimal;
-
 import ca.ulaval.glo4002.reservation.domain.fullcourse.Ingredient;
 import ca.ulaval.glo4002.reservation.domain.fullcourse.IngredientName;
 import ca.ulaval.glo4002.reservation.domain.fullcourse.Recipe;
 
 public class AllergyMain extends Recipe {
-  private static final BigDecimal TOFU_QUANTITY = BigDecimal.valueOf(10);
+  private static final double TOFU_QUANTITY = 10;
 
   public AllergyMain() {
     ingredients.add(new Ingredient(IngredientName.TOFU, TOFU_QUANTITY));
