+++ reservation-api/src/main/java/ca/ulaval/glo4002/reservation/domain/fullcourse/recipe/vegetarian/VegetarianStarter.java
@@ -1,14 +1,12 @@
 package ca.ulaval.glo4002.reservation.domain.fullcourse.recipe.vegetarian;
 
-import java.math.BigDecimal;
-
 import ca.ulaval.glo4002.reservation.domain.fullcourse.Ingredient;
 import ca.ulaval.glo4002.reservation.domain.fullcourse.IngredientName;
 import ca.ulaval.glo4002.reservation.domain.fullcourse.Recipe;
 
 public class VegetarianStarter extends Recipe {
-  private static final BigDecimal PUMPKIN_QUANTITY = BigDecimal.valueOf(5);
-  private static final BigDecimal CHOCOLATE_QUANTITY = BigDecimal.valueOf(8);
+  private static final double PUMPKIN_QUANTITY = 5;
+  private static final double CHOCOLATE_QUANTITY = 8;
 
   public VegetarianStarter() {
     ingredients.add(new Ingredient(IngredientName.PUMPKIN, PUMPKIN_QUANTITY));
